ai_scout_batch_2026_04_04/grounded-token-data-generator/requirements.txt
python-dotenv
requests
openai
transformers
sentence-transformers
scikit-learn
spacy
pandas
fastapi
uvicorn[standard]
pydantic